refactor: orders 조회 책임 이동
