+++ src/main/java/cart/domain/OrdersTaker.java
@@ -1,16 +1,12 @@
 package cart.domain;
 
-import cart.dto.OrdersResponse;
 import cart.exception.OrdersException;
 import cart.repository.CartItemRepository;
 import cart.repository.CouponRepository;
 import cart.repository.OrdersRepository;
-import cart.repository.ProductRepository;
 import org.springframework.stereotype.Component;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Component
@@ -18,13 +14,11 @@ public class OrdersTaker {
     private final OrdersRepository ordersRepository;
     private final CartItemRepository cartItemRepository;
     private final CouponRepository couponRepository;
-    private final ProductRepository productRepository;
 
-    public OrdersTaker(OrdersRepository ordersRepository, CartItemRepository cartItemRepository, CouponRepository couponRepository, ProductRepository productRepository) {
+    public OrdersTaker(OrdersRepository ordersRepository, CartItemRepository cartItemRepository, CouponRepository couponRepository) {
         this.ordersRepository = ordersRepository;
         this.cartItemRepository = cartItemRepository;
         this.couponRepository = couponRepository;
-        this.productRepository = productRepository;
     }
 
     public long takeOrder(final long memberId, final List<Long> cartIds, final List<Long> coupons) {
@@ -43,10 +37,6 @@ public class OrdersTaker {
                 .reduce(0, Integer::sum);
     }
 
-    private int calculateOrdersItemOriginalPrice(final long orderId) {
-        return productRepository.findProductIdQuantityWithOrdersId(orderId);
-    }
-
     private int calculateDiscountPrice(final long memberId, final int originalPrice, final List<Long> couponIds) {
         int discountPrice = originalPrice;
         List<Coupon> coupons = couponIds.stream()
@@ -61,30 +51,6 @@ public class OrdersTaker {
         return discountPrice;
     }
 
-    public List<OrdersResponse> findOrdersWithMember(final Member member) {
-        return reverse(ordersRepository.findAllOrdersByMember(member).stream()
-                .map(orders -> makeResponse(member, Optional.ofNullable(orders)))
-                .collect(Collectors.toList()));
-    }
-
-    public OrdersResponse findOrdersWithId(final Member member, final long id) {
-        return makeResponse(member, ordersRepository.findOrdersById(id));
-    }
-
-    private OrdersResponse makeResponse(final Member member, final Optional<Orders> orders) {
-        if (orders.isEmpty()) {
-            return OrdersResponse.noOrdersMembersResponse();
-        }
-        List<CartItem> ordersItem = ordersRepository.findCartItemByOrdersIds(member, orders.get().getId());
-        List<Coupon> coupons = couponRepository.findByOrdersId(orders.get().getId());
-        return OrdersResponse.of(
-                orders.get(),
-                ordersItem,
-                calculateOrdersItemOriginalPrice(orders.get().getId()),
-                coupons
-        );
-    }
-
     private void validateLimit(final int price, final int limit) {
         if (price < limit) {
             throw new OrdersException(limit + "이상으로 구매하셔야 합니다. ( 현재금액 : " + price + ")");
@@ -95,10 +61,5 @@ public class OrdersTaker {
         if (price < 0) {
             throw new OrdersException("금액은 음수가 될 수 없습니다.");
         }
-    }
-
-    private List<OrdersResponse> reverse(List<OrdersResponse> responses) {
-        Collections.reverse(responses);
-        return responses;
     }
 }
